requirements.txt
Django>=4.2,<5.0
python-dotenv==1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
whitenoise>=6.5.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0